+++ src/bencode.h
@@ -15,7 +15,7 @@
 
 
 int bc_decode_type(char c);
-int bc_read_string(char *in, char **str);
+size_t bc_read_string(char *in, char **str);
 int bc_read_int(char *in, int *val);
 
 #endif
